django_crud_app/django_crud_app/shops/serializer.py
from rest_framework import serializers
from .models import Shop


class ShopSerializer(serializers.ModelSerializer):
    class Meta:
        model = Shop
        fields = ('pk', 'name', 'shop_number', 'owner')
