django_crud_app/requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
defusedxml==0.6.0
Django==3.0.4
django-admin-tools==0.9.0
django-allauth==0.41.0
django-cors-headers==3.2.1
django-crispy-forms==1.9.0
django-rest-auth==0.9.5
django-rest-framework==0.1.0
django-webpack-loader==0.7.0
django-widget-tweaks==1.4.6
djangorestframework==3.11.0
GDAL==2.4.4
idna==2.9
mysqlclient==1.4.6
numpy==1.18.1
oauthlib==3.1.0
protobuf==3.11.4
python3-openid==3.1.0
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.8
